+++ docs/source/Changelog.rst
@@ -60,6 +60,10 @@ Added
    -  ``bom`` configuration
    -  ``kiri`` configuration
 
+-  PcbDraw:
+
+   -  ``show_copper`` to disable tracks and zones (from 1.2.0)
+
 -  Render 3D:
 
    -  KiCad 10: support for kicad-cli interface, adds some new options